refactor: address review on ScheduleActionCall dedup (#5386)
- Complete pending action requests by selecting matching ids first (narrowed
  by the indexed app_installation_id/state/type columns) and updating by
  primary key, instead of issuing the UPDATE with the jsonb predicate.
- Move the dedup repro and no-duplicate-on-retry tests into the existing
  integration_request_worker_test.go instead of a separate file.
