+++ pkg/models/integration_request.go
@@ -125,17 +125,35 @@ func (r *IntegrationRequest) Complete(tx *gorm.DB) error {
 // or self-rescheduling chains of the same action (#5386). Parameters are compared
 // as jsonb, which is canonical and key-order-independent, against the same
 // json.Marshal representation used when the request was created.
+//
+// Matching rows are selected first (narrowed by the indexed
+// app_installation_id/state/type columns) and then completed by primary key, so
+// the write touches only the few matched rows rather than scanning on the jsonb
+// predicate.
 func CompletePendingActionRequestsInTransaction(tx *gorm.DB, installationID uuid.UUID, actionName string, parameters any) error {
 	params, err := json.Marshal(parameters)
 	if err != nil {
 		return err
 	}
 
-	return tx.Model(&IntegrationRequest{}).
+	var ids []uuid.UUID
+	err = tx.Model(&IntegrationRequest{}).
 		Where("app_installation_id = ? AND state = ? AND type = ?",
 			installationID, IntegrationRequestStatePending, IntegrationRequestTypeInvokeAction).
 		Where("spec->'invoke_action'->>'action_name' = ?", actionName).
 		Where("spec->'invoke_action'->'parameters' = ?::jsonb", string(params)).
+		Pluck("id", &ids).
+		Error
+	if err != nil {
+		return err
+	}
+
+	if len(ids) == 0 {
+		return nil
+	}
+
+	return tx.Model(&IntegrationRequest{}).
+		Where("id IN ?", ids).
 		Updates(map[string]any{
 			"state":      IntegrationRequestStateCompleted,
 			"updated_at": time.Now(),
